% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizationIntervals.R
\name{asCategorizationIntervals}
\alias{asCategorizationIntervals}
\title{Converts a data.table to a CategorizationIntervals object}
\usage{
asCategorizationIntervals(dt_intervals)
}
\arguments{
\item{dt_intervals}{data.table must have columns value, min, max, close_left, close_right as defined in \code{\link{CategorizationIntervals}}.}
}
\value{
a CategorizationIntervals object.
}
\description{
Converts a data.table to a CategorizationIntervals object
}
\examples{
library(data.table)
dt_intervals <- data.table(value = c(1,2,3),
                           min = c(-Inf, 0, 0),
                           max = c(0 , 0, Inf),
                           close_left = c(0,1,0),
                           close_right = c(0,1,0) )
ci_intervals <- asCategorizationIntervals(dt_intervals)
print(ci_intervals)

}
\seealso{
\code{\link{CategorizationIntervals}}
}

