% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalizeFirstLetter.R
\name{capitalizeFirstLetter}
\alias{capitalizeFirstLetter}
\title{Capitalizes the first letter of each string in a character vector}
\usage{
capitalizeFirstLetter(v_s_string)
}
\arguments{
\item{v_s_string}{a character vector}
}
\value{
v_s_string with first letter capitalized of each string
}
\description{
Capitalizes the first letter of each string in a character vector
}
\examples{
capitalizeFirstLetter("this sentence should have the first letter capitalized.")
capitalizeFirstLetter(c("sentence one.", "sentence two"))
capitalizeFirstLetter("NO EFFECT")
capitalizeFirstLetter("a")
}

