% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategorizationIntervals.R
\name{CategorizationIntervals}
\alias{CategorizationIntervals}
\title{Construction function for CategorizationIntervals class}
\usage{
CategorizationIntervals(value, min, max, close_left, close_right,
  i_digits_label = 3)
}
\arguments{
\item{value}{a vector.}

\item{min}{a numeric vector.}

\item{max}{a numeric vector.}

\item{close_left}{a integer vector with only values 0 or 1 or a logical vector.}

\item{close_right}{a integer vector with only values 0 or 1 or a logical vector.}

\item{i_digits_label}{number of digits in label created as part of the object. Default is 3.}
}
\value{
a CategorizationIntervals object.
}
\description{
Construction function for CategorizationIntervals class used in \code{\link{categorizeByIntervals}}.
}
\details{
Intervals are not allowed to overlap. CategorizationIntervals class inherits from data.table. This object type is used in \code{\link{categorizeByIntervals}}.
}
\examples{
library(data.table)
ci_intervals <- CategorizationIntervals(value = c(1,2,3),
                                        min = c(-Inf, 0, 0),
                                        max = c(0 , 0, Inf),
                                        close_left = c(0,1,0),
                                        close_right = c(0,1,0) )
print(ci_intervals)
}
\seealso{
\code{\link{categorizeByIntervals}}
}

