% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorizeByIntervals.R
\name{categorizeByIntervals}
\alias{categorizeByIntervals}
\title{Categorize by intervals}
\usage{
categorizeByIntervals(v_n_values, ci_intervals)
}
\arguments{
\item{v_n_values}{a numeric vector. Must have length of at least 1.}

\item{ci_intervals}{a CategorizationIntervals object. See \code{\link{CategorizationIntervals}}.}
}
\value{
a vector with the same length as v_n_values with the same type as ci_intervals[['value']].
}
\description{
Categorizes numeric values allowing for different types of intervals.
}
\details{
Flexibles intervals are allowed in ci_intervals for example '(-Inf,0), [0,0], (0,Inf)' is valid.
}
\examples{
ci_intervals <- CategorizationIntervals(value = c(-1,0,1),
                                        min = c(-Inf, 0, 0),
                                        max = c(0 , 0, Inf),
                                        close_left = c(0,1,0),
                                        close_right = c(0,1,0) )
v_n_values <- -3:3
categorizeByIntervals(v_n_values, ci_intervals)

ci_intervals <- CategorizationIntervals(value = c("Neg","Zero","Pos"),
                                        min = c(-Inf, 0, 0),
                                        max = c(0 , 0, Inf),
                                        close_left = c(0,1,0),
                                        close_right = c(0,1,0) )
v_n_values <- -3:3
categorizeByIntervals(v_n_values, ci_intervals)
}
\seealso{
\code{\link{CategorizationIntervals}}
}

