% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CategorizationIntervals.R
\name{checkRepCategorizationIntervals}
\alias{checkRepCategorizationIntervals}
\title{Check rep of CategorizationIntervals}
\usage{
checkRepCategorizationIntervals(min = NULL, max = NULL, close_left = NULL,
  close_right = NULL)
}
\arguments{
\item{min}{a numeric vector.}

\item{max}{a numeric vector.}

\item{close_left}{a integer vector with only values 0 or 1 or a logical vector.}

\item{close_right}{a integer vector with only values 0 or 1 or a logical vector.}
}
\description{
Check rep of CategorizationIntervals
}
\keyword{internal}

