% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collectGarbage.R
\name{collectGarbage}
\alias{collectGarbage}
\title{Performs garbadge collections non verbosely}
\usage{
collectGarbage()
}
\value{
taken from  \code{\link{gc}}: gc returns a matrix with rows "Ncells"
}
\description{
Performs garbadge collections non verbosely
}
\examples{
collectGarbage()
}
\seealso{
\code{\link{gc}}
}

