% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countCharOccurrences.R
\name{countCharOccurrences}
\alias{countCharOccurrences}
\title{Counts the number of times a single characte occurs.}
\usage{
countCharOccurrences(char, v_s_string, b_ignore_case = FALSE)
}
\arguments{
\item{char}{a single character cannot be a regular expression.}

\item{v_s_string}{a character vector}

\item{b_ignore_case}{a boolean default FALSE of we want to ignore the case}
}
\value{
a integer vector of the number of times the character occurs in each string in v_s_string.
}
\description{
Counts the number of times a single characte occurs in each string of a character vector.
}
\examples{
countCharOccurrences("a", c("a other a", "zero"))
countCharOccurrences("a", c("A sentence."))
countCharOccurrences("a", c("A sentence."), b_ignore_case = TRUE)
}

