% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Frequency table of the number of times of individual value occurs}
\usage{
freqTable(dt_data, v_s_vars, b_total_row = FALSE, b_include_perc = FALSE)
}
\arguments{
\item{dt_data}{a data.table.}

\item{v_s_vars}{a character vector of column names in dt_data.}

\item{b_total_row}{default FALSE, whether we include a row with the total amount of values.}

\item{b_include_perc}{default FALSE, whether we include a column with percentages.}
}
\value{
A data.table with the number of times an individual value occurs
}
\description{
Frequency table of the number of times of individual value occurs
}
\examples{
library(data.table)
dt_data <- data.table( var1 = c(1,1,2,2,3,3,3,3),
                       var2 = c(6,6,6,7,7,7,7,7))
freqTable(dt_data, "var1", b_total_row = T)
freqTable(dt_data, c("var1","var2"), b_include_perc = T)
}

