% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepeatedRowsByKeys.R
\name{getRepeatedRowsByKeys}
\alias{getRepeatedRowsByKeys}
\title{Get repeated rows by key}
\usage{
getRepeatedRowsByKeys(dt_data, v_s_keys = NULL)
}
\arguments{
\item{dt_data}{a data.table.}

\item{v_s_keys}{a character vector of column names in \code{dt_data}.}
}
\value{
a data.table of all the rows which the combination of the values of \code{v_s_keys} are repeated. See \code{details}.
}
\description{
Returns a data.table of the repeated rows for a given key.
}
\details{
all rows of \code{dt_data} for which the combination of values of \code{v_s_keys} are repeated is returned. Additionally
a new column \code{rep_count_} is added  which indicates the number of times the key combination is repeated in the data set.
If there is no repeated cases function returns \code{NULL}.
}
\examples{
dt_data <- data.table( unique_identifier = c(1,2,3,3,3,4,4,5), value = rnorm(8))

getRepeatedRowsByKeys(dt_data, "unique_identifier")

}

