% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isColsUniqueIdentifier.R
\name{isColsUniqueIdentifier}
\alias{isColsUniqueIdentifier}
\title{Indicates if given fields are unique identifiers of a data.table}
\usage{
isColsUniqueIdentifier(dt_data, v_s_cols = NULL)
}
\arguments{
\item{dt_data}{a data.table.}

\item{v_s_cols}{a character vector of column names in dt_data which we expect allow for unique idenfication.}
}
\value{
A boolean of whether values in v_s_keys allow for unique idenfication of rows in dt_data.
}
\description{
Indicates if given fields allow for unique identification of rows i.e. if they are keys of the data.table.
}
\examples{
dt_data <- data.table(var1 = c(1,1,2,2) , var2 = c("a", "b" , "a","b") )
isColsUniqueIdentifier(dt_data, v_s_cols = "var1")
isColsUniqueIdentifier(dt_data, v_s_cols = c("var1", "var2"))
}

