% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isMonotonic.R
\name{isMonotonic}
\alias{isMonotonic}
\title{Indicates if a sequence of values are monotonic}
\usage{
isMonotonic(v_n_values, s_direction = "either")
}
\arguments{
\item{v_n_values}{numeric vector. Length of \code{v_n_values} must be at least 1.}

\item{s_direction}{a string of the monotonic direction we are interested. Must be 'decreasing' 'increasing' or 'either'.}
}
\value{
a boolean (logic vector length one) indicating if the sequence of values is monotonic. If v_n_values has a \code{NA} value the result is \code{NA}.
}
\description{
Indicates if a sequence of values are (non strictly) montonic increasing, decreasing or either.
}
\examples{
isMonotonic(c(1,1,2,3))
isMonotonic(c(1,1,NA, 2,3))
isMonotonic(1:5, s_direction = "increasing")
isMonotonic(1:5, s_direction = "decreasing")
isMonotonic(5:1, s_direction = "decreasing")
}

