% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeWithColPrioritization.R
\name{mergeWithColPrioritization}
\alias{mergeWithColPrioritization}
\title{Merger of two data.tables with repeated columns with prioritization}
\usage{
mergeWithColPrioritization(dt_prior, dt_other, v_s_keys)
}
\arguments{
\item{dt_prior}{data.table for which has prioritized columns.\code{v_s_keys} must be columns in \code{dt_prior}.}

\item{dt_other}{data.table wich we want to merge. \code{v_s_keys} must be columns in \code{dt_other}.}

\item{v_s_keys}{character vector of column names to merge by.}
}
\value{
data.table of prioritized merger of dt_prior and dt_other.
}
\description{
Merger (outer join) of two data.tables with repeated columns where we want to prioritize the values of one data.table.
}
\examples{
library(data.table)
i_num <- 3
dt_prior <- data.table( id = 1:i_num,
                        var1 = letters[1:i_num],
                        var2 = 111)
dt_other <- data.table( id = 1:(i_num+2),
                        var1 = paste0(letters[10 + 1:(i_num+2)],letters[10 + 1:(i_num+2)]) ,
                        var3 = -999)
v_s_keys <- "id"

dt_prior
dt_other
mergeWithColPrioritization(dt_prior, dt_other,v_s_keys )
}

