% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mode.R
\name{modeOfVector}
\alias{modeOfVector}
\title{Gives the mode of a vector}
\usage{
modeOfVector(v_values, b_allow_several_modes = TRUE,
  b_warning_if_several_modes = TRUE)
}
\arguments{
\item{v_values}{a vector.}

\item{b_allow_several_modes}{logical default TRUE. If we want to allow more than one mode.}

\item{b_warning_if_several_modes}{logical default TRUE. If we want a warning when there is more than one mode.}
}
\value{
A vector of length 1 with the mode of v_values. If there are more than one mode only one of the modes will be returned.
}
\description{
Gives the mode of a vector
}
\examples{
modeOfVector(1:10)
modeOfVector(c("a","a","b","c","c","c"))
}

