% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numUniqueValues.R
\name{numUniqueValues}
\alias{numUniqueValues}
\title{Gives the number of unique values of a vector}
\usage{
numUniqueValues(v_values)
}
\arguments{
\item{v_values}{a vector.}
}
\value{
An integer of the number of unique que values in v_values. If v_value has length zero returns 0.
}
\description{
Gives the number of unique values of a vector
}
\examples{
numUniqueValues(1:10)
numUniqueValues(c("a","a","b","c","c","c"))
}

