% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mode.R
\name{propCasesAreMode}
\alias{propCasesAreMode}
\title{Gives the proportion of cases which are the mode}
\usage{
propCasesAreMode(v_values)
}
\arguments{
\item{v_values}{a vector.}
}
\value{
A numeric of the proportion of cases which are the mode. If length(v_values) == 0 returns 0.
}
\description{
Gives the proportion of cases which are the mode
}
\examples{
propCasesAreMode(1:10)
propCasesAreMode(c("a","a","b","c","c","c"))
propCasesAreMode(numeric())
}

