% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrMakeObj.R
\name{readOrMakeObj}
\alias{readOrMakeObj}
\title{Read object from file or make object}
\usage{
readOrMakeObj(s_file, makeObj = NULL, ..., b_always_make_obj = FALSE,
  b_message_time = FALSE)
}
\arguments{
\item{s_file}{string of file where object is retreaved or saved see details. Must end in .rds.}

\item{makeObj}{function.}

\item{...}{additional parameters to be passed to \code{makeObj}.}

\item{b_always_make_obj}{boolean if true \code{makeObj} is called regardless of if \code{s_file} exists.}

\item{b_message_time}{boolean, should time ellapsed by messaged.}
}
\value{
either the object extracted from \code{s_file} or created by \code{makeObj}. See details.
}
\description{
Read object from file or make object
}
\details{
It is common to produce large objects which take a long time to execute as part of a script.
In some cases we would like to retreave these objects but in other cases they need to be constructed again.
This function helps with that processes. If s_file exist and \code{b_always_make_obj} is false object will be extracted.
If s_file does not exist or \code{b_always_make_obj} object will be created and saved to \code{s_file}.
}
\examples{
# create a random vector, assign it to v_n_random and save it to file.rds
v_n_random <- readOrMakeObj("file.rds", rnorm, n = 1000000, b_always_make_obj = TRUE, b_message_time = TRUE )

# retreave vector changing a parameter
v_n_random_2 <- readOrMakeObj("file.rds", rnorm, n = 100, b_always_make_obj = FALSE, b_message_time = TRUE )
length(v_n_random_2)
# 1000000 not 100
}

