% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizationIntervals.R
\name{CategorizationIntervals}
\alias{CategorizationIntervals}
\title{Construction function for CategorizationIntervals class}
\usage{
CategorizationIntervals(value, v_s_intervals = NULL, min = NULL,
  max = NULL, close_left = NULL, close_right = NULL, i_digits_label = 3)
}
\arguments{
\item{value}{a vector.}

\item{v_s_intervals}{a character vector. Each element must start with "[" or "(" and end with "]" or ")" and have two numeric values seperated by ",". Default NULL.}

\item{min}{a numeric vector. Default NULL.}

\item{max}{a numeric vector. Default NULL.}

\item{close_left}{a integer vector with only values 0 or 1 or a logical vector. Default NULL.}

\item{close_right}{a integer vector with only values 0 or 1 or a logical vector. Default NULL.}

\item{i_digits_label}{number of digits in label created as part of the object. Default is 3.}
}
\value{
a CategorizationIntervals object.
}
\description{
Construction function for CategorizationIntervals class used in \code{\link{categorizeByIntervals}}.
}
\details{
Intervals are not allowed to overlap. Either v_s_intervals is provided or min, max, close_left and close_right must be provided.  CategorizationIntervals class inherits from data.table. This object type is used in \code{\link{categorizeByIntervals}}.
}
\examples{
library(data.table)
ci_intervals <- CategorizationIntervals(value = c(1,2,3),
                                        min = c(-Inf, 0, 0),
                                        max = c(0 , 0, Inf),
                                        close_left = c(0,1,0),
                                        close_right = c(0,1,0) )
print(ci_intervals)
ci_intervals <- CategorizationIntervals(value = c(1,2,3),
                                        v_s_intervals = c("(-Inf,0)","[0,0]","(0,Inf)"))
print(ci_intervals)
}
\seealso{
\code{\link{categorizeByIntervals}}
}

