% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Frequency table of the number of times of individual value occurs}
\usage{
freqTable(dt_data, v_s_vars, b_total_row = FALSE, b_include_perc = FALSE,
  s_order_by = "v_s_vars")
}
\arguments{
\item{dt_data}{a data.table.}

\item{v_s_vars}{a character vector of column names in dt_data.}

\item{b_total_row}{default FALSE, whether we include a row with the total amount of values.}

\item{b_include_perc}{default FALSE, whether we include a column with percentages.}

\item{s_order_by}{default "v_s_vars". How the output data.table should be ordered. If "v_s_vars", ordered by values of parameter. If "ascending", ordered by ascending frequency. If "descending", ordered by "descending" frequency.}
}
\value{
A data.table with the number of times an individual value occurs
}
\description{
Frequency table of the number of times of individual value occurs
}
\examples{
library(data.table)
dt_data <- data.table( var1 = c(1,1,2,2,3,3,3,3),
                       var2 = c(6,6,6,7,7,7,7,7))
freqTable(dt_data, "var1", b_total_row = TRUE)
freqTable(dt_data, c("var1","var2"), b_include_perc = TRUE)
freqTable(dt_data, c("var1","var2"), b_include_perc = TRUE, s_order_by = "descending" )
}

